Relax type definition of `Oridnary*` AOs to allow arbitrary objects
In the spec these methods take "an Object", and not "an ordinary object".
They are indeed used on exotic objects, which might not have a
[[Prototype]] or [[IsExtensible]] internal slot, so the current signature
is wrong. I couldn't update the signature for the AOs related to
PrototypeOf and IsExtensible, because they don't actually work on
arbitrary objects even though the spec signature says so
(see tc39/ecma262#3612).

Thanks to this change, we can remove some internal slot from the
`ModuleNamespaceObject` interface, aligning it with the spec.